politician-copy-trading/requirements.txt
alpaca-py>=0.30.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
httpx>=0.27.0
pandas>=2.2.0
yfinance>=0.2.45
pytz>=2024.1
